Changes to be committed: 	renamed:    activity-tracking/Images/Confidential.png -> screenshot_feature/Images/Confidential.png 	renamed:    activity-tracking/Images/Vira3.png -> screenshot_feature/Images/Vira3.png 	renamed:    activity-tracking/Images/Viraico.png -> screenshot_feature/Images/Viraico.png 	renamed:    activity-tracking/Vira.kv -> screenshot_feature/Vira.kv 	renamed:    activity-tracking/Vira.py -> screenshot_feature/Vira.py 	modified:   screenshot_feature/screenshot.py
